refactor: renamed runTestsWithConfig to run-tests-with-config.js
